requirements.txt
# Core app framework
streamlit>=1.36.0

# Numerical & vector search
numpy>=1.24.0
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.6.1

# LLM API
groq>=0.5.0

# File parsing
PyPDF2>=3.0.1
pandas>=2.2.0
python-docx>=1.1.0
python-pptx>=0.6.23
markdown>=3.5.2

# Env configuration
python-dotenv>=1.0.1